Certain Security Changes
Added logout functionality and logged-out HTML page

Implemented the logout functionality for both teachers and students. When a user logs out, they are redirected to a logged-out HTML page ('logged_out.html'). This page displays a message informing the user that they have been logged out and prompts them to log in again to continue accessing the site. Also, added logic to restrict access to certain pages when the user is logged out, displaying an appropriate message if they attempt to access restricted pages.